requirements.txt
langchain
python-dotenv
ipykernel
langchain_community
pypdf
beautifulsoup4
streamlit
langchain_core
tensorflow
gradio
langchain-groq
fastapi
uvicorn
langserve
sse_starlette
langchain_chroma
langchain_huggingface
sentence-transformers
tf-keras
faiss-cpu
langchain_openai
arxiv
wikipedia
duckduckgo-search
sqlalchemy
langchain-sqlserver
pyodbc
validators
youtube-transcript-api
pytube
unstructured
numexpr
twilio
python-multipart
websockets
supabase